레벨 디자인 및 버그 수정

+++ Assets/Scripts/Interface/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void OnDamage(int _damage);
+    void OnDamage(float _damage);
 }
